pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context-evaluator"          # distribution name (pip install my-module)
description = "LMArena style environment to evaluate the LLMs ability to converse about a video"
license = { text = "MIT" }
readme = "README.md"
authors = []

version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.50.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "tqdm>=4.65.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
